Move conversion methods to common utilities file.
